[TA] Disable data port overflow test
In the interface `if_OS_Storage` data is exchanged via data ports which have a
limited size (PAGE_SIZE by default), so to make the component more robust all
the calls which are accessing this data port must check the requested size.

As currently guards are not implemented, the test is temporarily disabled.

See also:
SEOS-1733