[DOC] Document sizes used in this test system
RAM disk used in this test system is an underlying for the storage server.

As storage server is currently having three clients, RAM disk size must be
three times larger, as the size of each of the clients. Furthermore each client
must be large enough to store three test data arrays.