[FIX] Fix priorities workaround
Turns out that there are different `priority` variables in
CAmkES i.e.:

1. `_priority` - To adjust the priority of the control thread (the
                 thread that calls run).
2. `<interface>_priority` - To adjust the priority of an interface
                            thread.
3. `priority` - To adjust the priority of every thread within a
                given component instance.

Previously, and wrongly, `_priority` was set to low value, what does
not change the priority of the RPC calls, so in reality everything
still had the same priority, and tests were sporadically failing.

From now on, `priority` is set to low, so specific components' output
will not collide anymore with the SysLogger's output.

Note that the real fix for this `priority` workaround is in the
ticket SEOS-3131.